ENH: Add UNet nchan option to match Up out with Down out
The existing architecture has the same number of channels for both
the skip connection input and the input from the previous layer
(which needs to be upsampled) to the Up step.

We add the option to have the number of Up step output channels
match the number of channels in the corresponding Down step instead
(which is the number of channels in the skip connection input).