ENH: Add option to only have skip connection when downsampling
The existing method was to have a skip connection for every
block, which was not the intended behaviour.

We introduce a new parameter which can control whether to use
the previous method (always have a skip connection, even if
the unet block doesn't contain a downsampling step), and the
intended behaviour which is to have skip connection only
when the unet block includes a downsample and upsample.