BUG: Need to specify encoding when writing BOM
It turns out the EVL header opens with a Byte Order Marker (BOM), unicode
character \ufeff. This was included when I copied and pasted the header
from an existing EVL file. In order to write this on Windows, we need to
specify the encoding mode when we open the file.