ENH+API: Use context to open Echoview
Since a new Echoview window cannot be created if there is already one open,
we check to see if it is open first and if not open a window. If a new window
was opened, it is closed when leaving the context (it would close when the
application closes anyway). There are options for minimizing and hiding the
window while the code is running. A context is also used to open a file within
Echoview and automatically close it when finished. If the application is closed
when trying to open the file, Echoview is opened first and closed when the
context using the file is completed.

New CLI arguments are added to control whether the window is hidden or
minimized.