BUG: Fix line endings writen to evl/evr files on Windows
The \n was being automatically replaced with the OS-specific line ending, \r\n, resulting in our \r\n being turned into \r\r\n.
This is resolved by telling the file object which line ending we want to use (on all OSes) and letting it turn \n into that ending.